created js file for resources
