+++ index.html
@@ -3,7 +3,10 @@
 	<head>
 		<meta charset="UTF-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<meta
+			name="viewport"
+			content="width=device-width, initial-scale=1.0"
+		/>
 		<title>Resources API</title>
 		<!-- STYLESHEETS -->
 		<link rel="stylesheet" type="text/css" href="css/normalize.css" />
@@ -26,18 +29,26 @@
 				<section class="description">
 					<h1>Coding Resources API-</h1>
 					<p>
-						Search for coding resources by relevant keywords. This API serves educational content for a wide variety of computer science topics,
-						languages and technologies relevant to web development.
+						Search for coding resources by relevant keywords.
+						This API serves educational content for a wide
+						variety of computer science topics, languages and
+						technologies relevant to web development.
 					</p>
 				</section>
 				<section class="api-test">
 					<form action="">
-						<input type="submit" name="" value="" placeholder="/api/'keyword'" />
+						<input
+							type="submit"
+							name=""
+							value=""
+							placeholder="/api/'keyword'"
+						/>
 						<button id="keyword-btn" type="button">
 							<i class="fa-solid fa-magnifying-glass"></i>
 						</button>
 					</form>
-					<div class="result"></div>
+					<div class="result">
+					</div>
 				</section>
 			</main>
 			<section class="documentation">
