+++ resources.js
@@ -1,0 +1,19 @@
+const resources = [
+	{
+		name: 'Oh Shit, Git!',
+		url: 'https://ohshitgit.com/',
+		keywords: ['git', 'version control', 'command line'],
+	},
+	{
+		name: 'Javascript.info - Arrays',
+		url: 'https://javascript.info/array',
+		keywords: ['arrays'],
+	},
+	{
+		name: 'Building a Simple CRUD App with Node, Express, and MongoDB',
+		url: 'https://zellwk.com/blog/crud-express-mongodb/',
+		keywords: ['mongodb', 'express', 'node', 'backend'],
+	},
+];
+
+exports.resources = resources;
