+++ server.js
@@ -1,28 +1,11 @@
 const express = require('express');
 const app = express();
 const cors = require('cors');
+const { resources } = require("./resources");
 const PORT = process.env.PORT || 8000;
 app.use(cors());
 app.use(express.static(__dirname + '/public'));
 app.use('/public', express.static(__dirname + '/public'));
-
-const resources = [
-	{
-		name: 'Oh Shit, Git!',
-		url: 'https://ohshitgit.com/',
-		keywords: ['git', 'version control', 'command line'],
-	},
-	{
-		name: 'Javascript.info - Arrays',
-		url: 'https://javascript.info/array',
-		keywords: ['arrays'],
-	},
-	{
-		name: 'Building a Simple CRUD App with Node, Express, and MongoDB',
-		url: 'https://zellwk.com/blog/crud-express-mongodb/',
-		keywords: ['mongodb', 'express', 'node', 'backend'],
-	},
-];
 
 app.get('/', (req, res) => {
 	res.sendFile(__dirname + '/index.html');
